feat: Added new speckle exceptions and split into single files (#136)
* feat: Added new speckle exceptions and split into single files

* fix: Run format
